fix: rename app for SHA token generation
