+++ android/app/src/main/java/teamsite/rugby/MainActivity.java
@@ -1,4 +1,4 @@
-package io.ionic.starter;
+package teamsite.rugby;
 
 import com.getcapacitor.BridgeActivity;
