+++ android/app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <?xml version='1.0' encoding='utf-8'?>
 <resources>
-    <string name="app_name">TeamsiteApp</string>
-    <string name="title_activity_main">TeamsiteApp</string>
-    <string name="package_name">io.ionic.starter</string>
-    <string name="custom_url_scheme">io.ionic.starter</string>
+    <string name="app_name">teamsite-rugby-app</string>
+    <string name="title_activity_main">teamsite-rugby-app</string>
+    <string name="package_name">teamsite.rugby</string>
+    <string name="custom_url_scheme">teamsite.rugby</string>
 </resources>
